Reload form once content is saved, stops an exception
This is a hotfix for an issue where an exception within Twig was thrown because the form fields did not match up to the content after saving sometimes. This happened when adding a new repeatable group instances or removing an existing one.

The Field\Form class might want refactoring at some point to take this into account so it needn't generate the form twice.

Closes #170